+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 ## [Unreleased]
 ### Breaking Change
 * Heading classname changed from `.heading` to `.hogan-heading`
+* Deprecated `hogan/module/image/attachment/attr` filter in favor of `hogan/module/image/image/args`
+
+### Internal
+* Added filter `hogan/module/image/image/args` to filter image arguments